requirements.txt
flask==3.1.0
flask-cors==5.0.1
rembg==2.0.57
onnxruntime==1.21.1
Pillow==11.1.0
requests==2.32.3
